move cache into its own object so that we can start to cleanup logic
